+++ src/mx_sort_files.c
@@ -1,6 +1,6 @@
 #include "../inc/uls.h"
 
-void mx_sort_files(t_fileinfo files[], int size, int flags) {
+void mx_sort_files(t_fileinfo files[], int size, unsigned short flags) {
     bool (*cmp)(t_fileinfo, t_fileinfo) = NULL;
 
 	if(flags & FLAG_r) {
